Added our default wordPairs to AppDelegate so that they always are available
