+++ WordGame/WordGame/WordsViewController.swift
@@ -1,0 +1,27 @@
+//
+//  WordsViewController.swift
+//  WordGame
+//
+//  Created by Kseniia on 2025-03-26.
+//
+
+import UIKit
+
+class WordsViewController: UIViewController {
+
+    var wordPairs: [WordPair] = [
+    ]
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+            //if let encoded = try? JSONEncoder().encode(wordPairs) {
+                 
+                 //  UserDefaults.standard.set(encoded, forKey: "wordPairs")
+               //}
+   }
+
+
+}
+
